Updating build path: local and home paths
Local build is good for development and isolating
main .lore db file. Run with: `./build.sh local`.
It will build the db file in the root directory of
the project. Running with `home` or no args will
build the db file in the $HOME path if the file
doesnt already exist.